Support for logging client ip from X-Forwarded-For
This commit adds naïve support for extracting the real client from
the X-Forwarded-For header as set by various cache and proxy services.
This implementation does not enforce or require that upstream has set
X-Forwarded-For correctly, but merely splits on `, ` and is able to
interpolate %x in the logging format to the outermost clients ip.